// src/pdfs/ma/cjp27-petition-to-change-name-of-adult/schema.ts
/** Auto-generated from cjp27-petition-to-change-name-of-adult.pdf — do not edit */
import { PDFCheckBox, PDFTextField } from "@cantoo/pdf-lib";

export const pdfSchema = {
  oldFirstName: PDFTextField,
  oldMiddleName: PDFTextField,
  oldLastName: PDFTextField,
  birthplaceCity: PDFTextField,
  birthplaceState: PDFTextField,
  dateOfBirth: PDFTextField,
  residenceStreetAddress: PDFTextField,
  residenceCity: PDFTextField,
  residenceState: PDFTextField,
  residenceZipCode: PDFTextField,
  mailingStreetAddress: PDFTextField,
  mailingCity: PDFTextField,
  mailingState: PDFTextField,
  mailingZipCode: PDFTextField,
  email: PDFTextField,
  previousNameFrom: PDFTextField,
  previousNameTo: PDFTextField,
  previousNameReason: PDFTextField,
  otherNamesOrAliases: PDFTextField,
  newFirstName: PDFTextField,
  newMiddleName: PDFTextField,
  newLastName: PDFTextField,
  hasUsedOtherNameOrAliasFalse: PDFCheckBox,
  hasUsedOtherNameOrAliasTrue: PDFCheckBox,
  shouldReturnOriginalDocuments: PDFCheckBox,
  hasPreviousNameChangeFalse: PDFCheckBox,
  hasPreviousNameChangeTrue: PDFCheckBox,
  phoneNumber: PDFTextField,
  petitionerFirstName: PDFTextField,
  petitionerMiddleName: PDFTextField,
  petitionerLastName: PDFTextField,
  county: PDFTextField,
  isInterpreterNeeded: PDFCheckBox,
  isOkayToSharePronouns: PDFCheckBox,
  pronouns: PDFTextField,
  language: PDFTextField,
  reasonForChangingName: PDFTextField,
} as const;

export type PdfFieldName = keyof typeof pdfSchema;

export type PdfFieldValueType<T extends PdfFieldName> =
  (typeof pdfSchema)[T] extends typeof PDFCheckBox ? boolean : string;